Добавляет для формы авторизации новый компонент
