+++ src/components/authorization/form.jsx
@@ -1,0 +1,74 @@
+import React, {useRef} from "react";
+import {useHistory} from "react-router-dom";
+import PropTypes from "prop-types";
+import {connect} from "react-redux";
+import {login} from "src/store/api-actions";
+
+const AuthorizationForm = ({onSubmit}) => {
+  const loginRef = useRef();
+  const passwordRef = useRef();
+
+  const history = useHistory();
+
+  const handleSubmit = (evt) => {
+    evt.preventDefault();
+
+    onSubmit({
+      login: loginRef.current.value,
+      password: passwordRef.current.value,
+    });
+  };
+
+  return (
+    <form
+      className="login__form form"
+      action="#"
+      method="post"
+      onSubmit={handleSubmit}
+    >
+      <div className="login__input-wrapper form__input-wrapper">
+        <label className="visually-hidden">E-mail</label>
+        <input
+          ref={loginRef}
+          className="login__input form__input"
+          type="email"
+          name="email"
+          placeholder="Email"
+          required
+        />
+      </div>
+      <div className="login__input-wrapper form__input-wrapper">
+        <label className="visually-hidden">Password</label>
+        <input
+          ref={passwordRef}
+          className="login__input form__input"
+          type="password"
+          name="password"
+          placeholder="Password"
+          required
+        />
+      </div>
+      <button
+        onClick={() => history.push(`/`)}
+        className="login__submit form__submit button"
+        type="submit"
+      >
+        Sign in
+      </button>
+    </form>
+  );
+};
+
+AuthorizationForm.propTypes = {
+  onSubmit: PropTypes.func.isRequired,
+};
+
+
+const mapDispatchToProps = (dispatch) => ({
+  onSubmit(authData) {
+    dispatch(login(authData));
+  }
+});
+
+export {AuthorizationForm};
+export default connect(null, mapDispatchToProps)(AuthorizationForm);
