+++ src/components/pages/login-page/login-page.jsx
@@ -1,24 +1,8 @@
-import React, {useRef} from "react";
-import {useHistory} from "react-router-dom";
-import PropTypes from "prop-types";
-import {connect} from "react-redux";
-import {login} from "src/store/api-actions";
+import React from "react";
 import Header from "src/components/layout/header/header";
+import AuthorizationForm from "src/components/authorization/form";
 
-const LoginPage = ({onSubmit}) => {
-  const loginRef = useRef();
-  const passwordRef = useRef();
-
-  const history = useHistory();
-
-  const handleSubmit = (evt) => {
-    evt.preventDefault();
-
-    onSubmit({
-      login: loginRef.current.value,
-      password: passwordRef.current.value,
-    });
-  };
+const LoginPage = () => {
 
   return (
     <div className="page page--gray page--login">
@@ -27,42 +11,7 @@ const LoginPage = ({onSubmit}) => {
         <div className="page__login-container container">
           <section className="login">
             <h1 className="login__title">Sign in</h1>
-            <form
-              className="login__form form"
-              action="#"
-              method="post"
-              onSubmit={handleSubmit}
-            >
-              <div className="login__input-wrapper form__input-wrapper">
-                <label className="visually-hidden">E-mail</label>
-                <input
-                  ref={loginRef}
-                  className="login__input form__input"
-                  type="email"
-                  name="email"
-                  placeholder="Email"
-                  required
-                />
-              </div>
-              <div className="login__input-wrapper form__input-wrapper">
-                <label className="visually-hidden">Password</label>
-                <input
-                  ref={passwordRef}
-                  className="login__input form__input"
-                  type="password"
-                  name="password"
-                  placeholder="Password"
-                  required
-                />
-              </div>
-              <button
-                onClick={() => history.push(`/`)}
-                className="login__submit form__submit button"
-                type="submit"
-              >
-                Sign in
-              </button>
-            </form>
+            <AuthorizationForm />
           </section>
           <section className="locations locations--login locations--current">
             <div className="locations__item">
@@ -77,16 +26,4 @@ const LoginPage = ({onSubmit}) => {
   );
 };
 
-LoginPage.propTypes = {
-  onSubmit: PropTypes.func.isRequired,
-};
-
-
-const mapDispatchToProps = (dispatch) => ({
-  onSubmit(authData) {
-    dispatch(login(authData));
-  }
-});
-
-export {LoginPage};
-export default connect(null, mapDispatchToProps)(LoginPage);
+export default LoginPage;
